akyc_edge_docker/requirements.txt
paho-mqtt
redis
pyyaml
